requirements.txt
yfinance
pandas
numpy
ta
python-telegram-bot==20.3
python-dotenv
